moved MeasureTool component to a helper file
