+++ .vscode/settings.json
@@ -1,5 +1,10 @@
 {
     "git.ignoreLimitWarning": true,
     "editor.defaultFormatter": "esbenp.prettier-vscode",
-    "editor.formatOnSave": true
+    "editor.formatOnSave": true,
+    "workbench.colorCustomizations": {
+        "activityBar.background": "#023521",
+        "titleBar.activeBackground": "#024B2E",
+        "titleBar.activeForeground": "#EDFEF7"
+    }
 }
